Segment TrackPage UI into header and rec section
This makes the track page cleaner and makes it so that the audio
features and track info are loaded in first. The Recs request won't
go out until after the track info has loaded in. Also changed UX so
that that recs are automatically loaded on section load instead of
having to press the 'recommend' button. recs are the first thing you see
now; not the options.